Polynomial: fix bug with special funcs
When a special function should not be called with the type given,
it should return NotImplemented. This tells python to try the other
one. For example: t is a term and p is a polynomial. t + p calls
t.__add__(p), but we really want to call p.__radd__(t).